+++ src/components/Renderer/index.js
@@ -40,6 +40,7 @@ import {
 import {
   ContextProvider as FrontEditorContextProvider,
   SubscriptionProvider as FrontEditorSubscriptionProvider,
+  FrontEditorRoot,
 } from "@prisma-cms/front-editor"
 
 
@@ -60,110 +61,127 @@ export default class BoilerplateRenderer extends PrismaCmsRenderer {
 
     let routers = [
       {
-        exact: true,
-        path: "/users/:userId",
-        render: (props) => {
-          const {
-            params,
-          } = props.match;
-
-          const {
-            userId,
-          } = params || {};
-
-          return <UserPage
-            key={userId}
-            getQueryFragment={getQueryFragment}
-            where={{
-              id: userId,
-            }}
-            {...props}
-          />
-        }
-      },
-      {
-        exact: true,
+        exact: false,
         path: "/",
-        component: ChatRoomsPage,
-      },
-      {
-        exact: true,
-        path: "/chat-rooms",
-        component: ChatRoomsPage,
-      },
-      {
-        exact: true,
-        path: "/chat-rooms/create",
-        component: CreateChatRoomPage,
-      },
-      {
-        exact: true,
-        path: "/chat-rooms/:id",
-        render: props => {
-
-          const {
-            match: {
-              params: {
-                id,
-              },
-            },
-          } = props;
-
-          return <ChatRoomPage
-            key={id}
-            where={{
-              id,
-            }}
-            {...props}
-          />
-        },
-      },
-      {
-        exact: true,
-        path: "/chat-messages",
-        component: ChatMessagesPage,
-      },
-      {
-        exact: true,
-        path: "/chat-messages/:id",
-        render: props => {
-
-          const {
-            match: {
-              params: {
-                id,
-              },
-            },
-          } = props;
-
-          return <ChatMessagePage
-            key={id}
-            where={{
-              id,
-            }}
-            {...props}
-          />
-        },
-      },
-      {
-        exact: true,
-        path: "/eth-transactions",
-        render: props => <TransactionsPage
-          {...props}
-          where={{}}
-          first={10}
-          orderBy="createdAt_DESC"
-        />
-      },
-      {
-        exact: true,
-        path: "/eth-transactions/:transactionId",
-        component: TransactionPage,
-      },
-    ].concat(super.getRoutes());
+        component: FrontEditorRoot,
+      }
+    ];
 
     return routers;
   }
+
+  // getRoutes() {
+
+  //   const {
+  //     getQueryFragment,
+  //   } = this.context;
+
+  //   let routers = [
+  //     {
+  //       exact: true,
+  //       path: "/users/:userId",
+  //       render: (props) => {
+  //         const {
+  //           params,
+  //         } = props.match;
+
+  //         const {
+  //           userId,
+  //         } = params || {};
+
+  //         return <UserPage
+  //           key={userId}
+  //           getQueryFragment={getQueryFragment}
+  //           where={{
+  //             id: userId,
+  //           }}
+  //           {...props}
+  //         />
+  //       }
+  //     },
+  //     {
+  //       exact: true,
+  //       path: "/",
+  //       component: ChatRoomsPage,
+  //     },
+  //     {
+  //       exact: true,
+  //       path: "/chat-rooms",
+  //       component: ChatRoomsPage,
+  //     },
+  //     {
+  //       exact: true,
+  //       path: "/chat-rooms/create",
+  //       component: CreateChatRoomPage,
+  //     },
+  //     {
+  //       exact: true,
+  //       path: "/chat-rooms/:id",
+  //       render: props => {
+
+  //         const {
+  //           match: {
+  //             params: {
+  //               id,
+  //             },
+  //           },
+  //         } = props;
+
+  //         return <ChatRoomPage
+  //           key={id}
+  //           where={{
+  //             id,
+  //           }}
+  //           {...props}
+  //         />
+  //       },
+  //     },
+  //     {
+  //       exact: true,
+  //       path: "/chat-messages",
+  //       component: ChatMessagesPage,
+  //     },
+  //     {
+  //       exact: true,
+  //       path: "/chat-messages/:id",
+  //       render: props => {
+
+  //         const {
+  //           match: {
+  //             params: {
+  //               id,
+  //             },
+  //           },
+  //         } = props;
+
+  //         return <ChatMessagePage
+  //           key={id}
+  //           where={{
+  //             id,
+  //           }}
+  //           {...props}
+  //         />
+  //       },
+  //     },
+  //     {
+  //       exact: true,
+  //       path: "/eth-transactions",
+  //       render: props => <TransactionsPage
+  //         {...props}
+  //         where={{}}
+  //         first={10}
+  //         orderBy="createdAt_DESC"
+  //       />
+  //     },
+  //     {
+  //       exact: true,
+  //       path: "/eth-transactions/:transactionId",
+  //       component: TransactionPage,
+  //     },
+  //   ].concat(super.getRoutes());
+
+  //   return routers;
+  // }
 
 
   renderWrapper() {