+++ src/components/Renderer/pages/UsersPage/UserPage/index.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 
-import UserPageProto from "@prisma-cms/front/lib/modules/pages/UsersPage/UserPage";
+import UserPageProto from "@prisma-cms/front/lib/components/pages/UsersPage/UserPage";
 
 
 import View from "./View";