+++ src/components/Renderer/pages/layout/index.js
@@ -3,7 +3,7 @@ import React, { Component, Fragment } from 'react';
 
 import PropTypes from 'prop-types';
 
-import PrismaCmsPageLayout from "@prisma-cms/front/lib/modules/pages/layout";
+import PrismaCmsPageLayout from "@prisma-cms/front/lib/components/pages/layout";
 
 export default class PageLayout extends PrismaCmsPageLayout {